Update upgrade functionality
Updated fire rate and damage increase upgrades for all towers

+++ Tower-Tactics/Assets/Scripts/Tower/Laser/Laser.cs
@@ -9,7 +9,7 @@ public class Laser : MonoBehaviour
     public TowerAiming towerAiming;
     private GameObject spawnedLaser;
     private LineRenderer lineRenderer;
-    public float dps = 1f;
+    public int dps = 1;
     // Start is called before the first frame update
     void Start()
     {
